requirements.txt
backports.csv==1.0.7
defusedxml==0.5.0
diff-match-patch==20181111
dj-pagination==2.4.0
Django==2.2
django-import-export==1.2.0
et-xmlfile==1.0.1
gunicorn==19.9.0
jdcal==1.4
odfpy==1.4.0
openpyxl==2.6.2
pytz==2019.1
PyYAML==5.1
sqlparse==0.3.0
tablib==0.13.0
xlrd==1.2.0
xlwt==1.3.0
